+++ CHANGELOG.md
@@ -1,9 +1,9 @@
-#3.3.6
-- Improved dark background recognition on some websites
-
 #3.3.5
 - Fixed dublicated Translate button issue
 - Made cryptocompare.com permission explicit in manifest
+- Improved dark background recognition on some websites
+- Attempt to fix incompatibilities with older browsers
+- Improved 'amount of icons in row' search tooltip config
 
 #3.3.4
 - Implemented 'amount of icons in row' config